Move api/leaderboard.js into api/campaigns.js
